feat：优化重复代码
