+++ server/api/admin/songs/duration.post.ts
@@ -2,6 +2,7 @@ import { readBody } from 'h3'
 import { db } from '~/drizzle/db'
 import { songs } from '~/drizzle/schema'
 import { eq } from 'drizzle-orm'
+import { requireSongAdmin } from '~~/server/utils/requireSongAdmin'
 import { createApiError } from '~~/server/utils/apiError'
 import { SERVER_ERROR_CODES } from '~~/server/config/constants'
 import {
@@ -11,13 +12,7 @@ import {
 } from '~~/server/utils/native_tx'
 
 export default defineEventHandler(async (event) => {
-  const user = event.context.user
-  if (!user) {
-    throw createApiError(401, SERVER_ERROR_CODES.AUTH_UNAUTHORIZED, '未授权访问')
-  }
-  if (!['SONG_ADMIN', 'ADMIN', 'SUPER_ADMIN'].includes(user.role)) {
-    throw createApiError(403, SERVER_ERROR_CODES.COMMON_INSUFFICIENT_PERMISSION, '没有权限访问')
-  }
+  requireSongAdmin(event)
 
   const body = await readBody(event)
   const songId = body?.songId == null ? null : Number(body.songId)
